requirements.txt
asgiref==3.8.1
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
python-environ==0.4.54
sqlparse==0.5.3
tzdata==2025.1
whitenoise==6.11.0
django-filter
django-cors-headers
Pillow
djangorestframework-simplejwt
djangorestframework-simplejwt[blacklist]
drf-yasg


